Setup for Banking System
Made it so that each player has [playername]_money and the amount of
money he has in his bank on the server.
Made it so the payment system updates the bank stuff and not the money
on the player.
Moved defines over to a separate file to clean things up.
Fixed gametick so it actually works on both a dedi server and a testing
environment.